move revision tag integ test to pilot/revisions
